Fix wrong implementation of api for plugins : . plugin_perspective_camera . plugin_point_light . plugin_simple_material
