+++ src/Plugins/PerspectiveCamera/CMakeLists.txt
@@ -1,0 +1,7 @@
+cmake_minimum_required(VERSION 3.13)
+
+target_include_directories(plugin_perspective_camera PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+
+target_sources(plugin_perspective_camera PRIVATE
+        "${CMAKE_CURRENT_SOURCE_DIR}/PerspectiveCamera.cpp"
+        )
