+++ src/Plugins/PerspectiveCamera/PerspectiveCamera.cpp
@@ -1,14 +1,19 @@
 /*
 ** EPITECH PROJECT, 2023
-** Raytracer
+** raytracer
 ** File description:
-** PerspectiveCameraApi
+** PerspectiveCamera
 */
 
+#include <libconfig.h++>
 #include "PerspectiveCamera.hpp"
-#include "PerspectiveCameraApi.hpp"
 
-Raytracer::IEntity *Plugin::PerspectiveCameraApi::createEntity(const libconfig::Setting &setting) {
+Plugin::PerspectiveCamera::PerspectiveCamera(const Component::Vector3f &position, const Component::Vector3f &rotation, float field_of_view, float aspect_ratio)
+        : Raytracer::ACam("PerspectiveCamera", position, rotation, field_of_view, aspect_ratio)
+{
+}
+
+Raytracer::IEntity *createEntity(const libconfig::Setting &setting) {
     Component::Vector3f position(setting["position"][0], setting["position"][1], setting["position"][2]);
     Component::Vector3f rotation(setting["rotation"][0], setting["rotation"][1], setting["rotation"][2]);
     float field_of_view = setting["field_of_view"];
@@ -17,10 +22,10 @@ Raytracer::IEntity *Plugin::PerspectiveCameraApi::createEntity(const libconfig::
     return new Plugin::PerspectiveCamera(position, rotation, field_of_view, aspect_ratio);
 }
 
-const char *Plugin::PerspectiveCameraApi::getName() const {
+const char *getName() {
     return "PerspectiveCamera";
 }
 
-Raytracer::CompType Plugin::PerspectiveCameraApi::getType() const {
+Raytracer::CompType getType() {
     return Raytracer::CompType::CAM;
 }
