+++ src/Plugins/PointLight/CMakeLists.txt
@@ -1,9 +1,7 @@
 cmake_minimum_required(VERSION 3.13)
 
-target_include_directories(plugin_plane PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
-target_include_directories(raytracer PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+target_include_directories(plugin_point_light PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 
-target_sources(plugin_plane PRIVATE
-        "${CMAKE_CURRENT_SOURCE_DIR}/PointLight.cpp",
-        "${CMAKE_CURRENT_SOURCE_DIR}/PointLightApi.cpp"
+target_sources(plugin_point_light PRIVATE
+        "${CMAKE_CURRENT_SOURCE_DIR}/PointLight.cpp"
         )
