+++ src/Plugins/SimpleMaterial/CMakeLists.txt
@@ -1,9 +1,7 @@
 cmake_minimum_required(VERSION 3.13)
 
-target_include_directories(plugin_sphere PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
-target_include_directories(raytracer PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+target_include_directories(plugin_simple_material PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 
-target_sources(plugin_sphere PRIVATE
+target_sources(plugin_simple_material PRIVATE
         "${CMAKE_CURRENT_SOURCE_DIR}/SimpleMaterial.cpp"
-        "${CMAKE_CURRENT_SOURCE_DIR}/SimpleMaterialApi.cpp"
         )
