+++ src/Plugins/Sphere/Sphere.cpp
@@ -5,6 +5,8 @@
 ** Sphere
 */
 
+#include <iostream>
+#include <libconfig.h++>
 #include "Sphere.hpp"
 
 Plugin::Sphere::Sphere(const Component::Vector3f &position, float radius) : APrimitive("Sphere", position), _radius(radius)
@@ -40,4 +42,33 @@ float Plugin::Sphere::intersect(const Raytracer::Ray &ray) const
 
 Component::Vector3f Plugin::Sphere::getNormal(const Component::Vector3f &hit_point) const {
     return (hit_point - getPosition()).normalize();
+}
+
+Raytracer::IEntity *createEntity(const libconfig::Setting &setting) {
+    Component::Vector3f position;
+    float radius = 0;
+
+    try {
+        if (setting.exists("position")) {
+            position = Component::Vector3f(setting["position"][0],
+            setting["position"][1], setting["position"][2]);
+        } else {
+            position = Component::Vector3f(0, 0, 0);
+        }
+        if (setting.exists("radius")) {
+            radius = setting["radius"];
+        }
+    } catch (const libconfig::SettingException &e) {
+        std::cerr << e.what() << std::endl;
+        return nullptr;
+    }
+    return new Plugin::Sphere(position, radius);
+}
+
+const char *getName() {
+    return "Sphere";
+}
+
+Raytracer::CompType getType() {
+    return Raytracer::CompType::PRIMITIVE;
 }
