+++ src/Raytracer/Entity/CMakeLists.txt
@@ -3,6 +3,9 @@ cmake_minimum_required(VERSION 3.13)
 target_include_directories(raytracer PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 target_include_directories(plugin_sphere PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 target_include_directories(plugin_plane PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+target_include_directories(plugin_perspective_camera PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+target_include_directories(plugin_point_light PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
+target_include_directories(plugin_simple_material PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 
 target_sources(raytracer PRIVATE
         "${CMAKE_CURRENT_SOURCE_DIR}/ACam.cpp"
